Day_5/Vision_Challenge/requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
comm==0.2.3
confection==0.1.5
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
easyocr==1.7.2
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
filelock==3.20.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.13.2
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
lazy_loader==0.4
Levenshtein==0.27.3
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.3
numpy==2.3.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.2
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-hub==0.16.1
tensorflow-text==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
thinc==8.3.10
tifffile==2026.2.20
tokenizers==0.22.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
triton==3.3.1
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
Werkzeug==3.1.6
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
